# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ProductVariant:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        description: e.g., "Small - Red"
        type: string
      price:
        type: number
      product_id:
        type: integer
      size:
        type: string
      sku:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@karimastore.com
    name: Karima Store API Support
  description: |-
    Karima Store E-commerce API with JWT Authentication

    ## Authentication

    This API uses **JWT (JSON Web Token)** for authentication.

    ### For Web/Browser Clients:
    1. Login via /api/v1/auth/login endpoint
    2. Session cookie (jwt) will be set automatically
    3. Make API requests with the cookie included

    ### For API/Mobile Clients:
    1. Obtain JWT token from /api/v1/auth/login endpoint
    2. Include token in requests:
    - Method 1: Authorization: Bearer <jwt_token>
    - Method 2: X-Session-Token: <jwt_token> header

    ### Authorization Levels:
    - **Public**: No authentication required (GET endpoints for browsing)
    - **Authenticated**: Valid JWT token required
    - **Admin**: Valid JWT token + admin role
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Karima Store API
  version: "1.0"
paths:
...
  /api/v1/products/{product_id}/variants:
    get:
      description: Retrieve all variants for a specific product
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get variants by product ID
      tags:
      - variants
...
  /api/v1/variants:
    post:
      consumes:
      - application/json
      description: Create a new product variant with the provided details
      parameters:
      - description: Variant object
        in: body
        name: variant
        required: true
        schema:
          $ref: '#/definitions/models.ProductVariant'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new variant
      tags:
      - variants
  /api/v1/variants/{id}:
    delete:
      description: 'Delete a variant by ID. **Admin only**: Requires authentication
        with admin role.'
      parameters:
      - description: Variant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Delete variant (Admin only)
      tags:
      - variants
    get:
      description: Retrieve a single variant by its ID
      parameters:
      - description: Variant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get variant by ID
      tags:
      - variants
    put:
      consumes:
      - application/json
      description: 'Update an existing variant by ID. **Admin only**: Requires authentication
        with admin role.'
      parameters:
      - description: Variant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Variant object
        in: body
        name: variant
        required: true
        schema:
          $ref: '#/definitions/models.ProductVariant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Update variant (Admin only)
      tags:
      - variants
  /api/v1/variants/{id}/stock:
    patch:
      consumes:
      - application/json
      description: 'Update the stock quantity for a variant. **Admin only**: Requires
        authentication with admin role.'
      parameters:
      - description: Variant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Stock update object
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Update variant stock (Admin only)
      tags:
      - variants
  /api/v1/variants/sku/{sku}:
    get:
      description: Retrieve a single variant by its SKU
      parameters:
      - description: Variant SKU
        in: path
        name: sku
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get variant by SKU
      tags:
      - variants
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"